Level1 Train Puzzle Stuff(On going)
Includes Train Levers, Door, Lights for the puzzle inside the train.

For now the lights does not do anything due I'm not clear how the lights should behave for the puzzle.

Modified the prototypes from Matthew in order to implement it into the level.

Those I modified I added my initial, those that I haven't touched but required in order for it to work, I added Matthew's initial.